requirements.txt
aiogram==3.4.1
aiohttp
fastapi
python-dotenv
sqlite-utils
httpx
